Rename files and divided into two parts tests for automata
