+++ project/automata.py
